requirements.txt
bcrypt==3.2.0
coloredlogs==15.0
progressbar==2.5
tabulate==0.8.9
PyYAML==5.3.1
psutil==5.8.0
docstring-parser==0.7.3
mkdocstrings==0.15.2
marshmallow==3.11.1
pytest==6.2.2
python-crontab==2.5.1
python-daemon==2.3.0
requests==2.24.0
sqlalchemy==1.3.18
Unidecode==1.2.0